backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
supabase
openai
apscheduler
jinja2
parsel
scrapy
arxiv